fix : 핸들러 중복삭제
